// TrelloCloneWebUI/src/app/swagger/model/cardRequest.ts
/**
 * Trello Clone Core Api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface CardRequest { 
    id?: number;
    name?: string;
    priority?: number;
    boardId?: number;
}