// TrelloCloneWebUI/src/app/swagger/model/taskRequest.ts
/**
 * Trello Clone Core Api
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface TaskRequest { 
    priority?: number;
    createdDate?: Date;
    dueDate?: Date;
    cardId?: number;
    id?: number;
    name?: string;
    description?: string;
}